Gate system.mail asserts on schema content
system.mail is not FullyValidatable in Drupal ^10.4, so its type asserts
fail there. Move them to a separate fixture that only runs when the
installed core's schema actually has the constraint, keyed off the
schema file content rather than a version compare.
